score_service/requirements.txt
grpcio
protobuf
requests
paddleocr
paddlepaddle
pillow
numpy
python-dotenv
opencv-python
